+++ src/ui/BottomDrawer/SelectBottomDrawer/index.ts
@@ -1,0 +1,1 @@
+export { default as SelectBottomDrawer } from "./SelectBottomDrawer";
